A Colourful Life!
the player's colour changes on every collision matching the colour of the ruin; will need to add reward components and feedback to make the game feel more satisfying.

There are currently 4 colours (subject to change)
Will need to fine tune the spawn rate and to reconsider the shapes of the ruins too.